Do not "upload" the kit into kits/<version> when "angela.kitInstallationDir" is set. Angela should re-use as is the local kit provided by Gradle in order to leverage Gradle's ability to refresh the content when some changes happen in the projec. Note: This is only valid when "angela.kitInstallationDir" is used

+++ common/src/main/java/org/terracotta/angela/common/tcconfig/License.java
@@ -47,10 +47,12 @@ public class License {
 
   public File writeToFile(File dest) {
     File licenseFile = new File(dest, this.filename);
-    try {
-      Files.write(licenseFile.toPath(), licenseContent.getBytes());
-    } catch (IOException ioe) {
-      throw new RuntimeException(ioe);
+    if(!licenseFile.exists()) {
+      try {
+        Files.write(licenseFile.toPath(), licenseContent.getBytes());
+      } catch (IOException ioe) {
+        throw new RuntimeException(ioe);
+      }
     }
     return licenseFile;
   }
